fix(lint): resolve golangci-lint failures
Address all 33 issues reported by golangci-lint that were failing the
Lint workflow.

- errcheck: explicitly discard error returns on best-effort cleanup
  calls (FindObjectsFinal, Close, Logout, CloseSession, Finalize,
  DestroyObject, Destroy, os.RemoveAll), mostly in test setup/teardown
- revive: add doc comments to exported PublicKey/PrivateKey.Object
  methods; rename FindObjects' `max` parameter to `maxObjects` to avoid
  shadowing the builtin
- unconvert: drop redundant C.size_t() conversions where the value is
  already CK_ULONG
- add make lint

No behavior change. Verified with go build, go vet, and test compilation.

+++ cryptoki/marshal.go
@@ -42,7 +42,7 @@ func zfree(p unsafe.Pointer, n C.CK_ULONG) {
 		return
 	}
 	if n > 0 {
-		C.ck_memzero(p, C.size_t(n))
+		C.ck_memzero(p, n)
 	}
 	C.free(p)
 }
@@ -158,7 +158,7 @@ func outOp(call func(out C.CK_BYTE_PTR, outLen *C.CK_ULONG) C.CK_RV) ([]byte, er
 		if uint64(n) > maxOutBuf {
 			return nil, fmt.Errorf("cryptoki: output size %d exceeds limit %d", uint64(n), uint64(maxOutBuf))
 		}
-		buf := C.malloc(C.size_t(n))
+		buf := C.malloc(n)
 		rv := uint(call((C.CK_BYTE_PTR)(buf), &n))
 		if rv == CKR_BUFFER_TOO_SMALL {
 			// n now holds the larger required size; reallocate and retry.
